Make script work better for projects with submodules

+++ find-test-scope-problems.sh
@@ -5,10 +5,25 @@ if [ ! which xmlstarlet 2> /dev/null ] ; then
   exit 1
 fi
 
+function run {
+  "$@" &> out.tmp
+  if [ 0 -ne $? ] ; then
+    cat out.tmp
+    rm out.tmp
+    return 1
+  fi
+  rm out.tmp
+  return 0
+}
+
 function deps {
   local out=$1
   local in=$2
-  mvn -B dependency:list -DoutputFile=$out.raw -f $in &> /dev/null
+  set +e
+  if ! run mvn -B dependency:list -DoutputFile=$out.raw -f $in ; then
+    exit 1
+  fi
+  set -e
   cat $out.raw | grep "^ "| grep ":compile" | cut -d : -f 1-2 | sort > ${out}
 }
 
@@ -16,6 +31,9 @@ function check {
   echo "Running check for $(pwd):"
 
   if grep -F "<modules>" pom.xml > /dev/null ; then
+    # This is needed if the submodules depend on each other or the parent
+    run mvn -B -DskipTests install
+
     for file in $(find . | grep "/pom.xml$") ; do
       local dir=$(dirname $file)
       if [ $dir != "." ] ; then
@@ -25,7 +43,6 @@ function check {
       fi
     done
   else
-
     set -e
     xmlstarlet ed -P -N ns=http://maven.apache.org/POM/4.0.0 -d 'ns:project/ns:dependencies/ns:dependency[ns:scope="test"]' pom.xml > pom.notest.xml
     deps dependency-list.txt pom.xml
